requirements.txt
pyserial>=3.5
matplotlib>=3.5
numpy>=1.20
